+++ src/CRUDlex/CRUDEntityDefinition.php
@@ -58,6 +58,11 @@ class CRUDEntityDefinition {
     protected $deleteCascade;
 
     /**
+     * The amount of items to display per page on the listview.
+     */
+    protected $pageSize;
+
+    /**
      * Holds the {@see CRUDServiceProvider}.
      */
     protected $serviceProvider;
@@ -156,10 +161,12 @@ class CRUDEntityDefinition {
      * field names
      * @param boolean $deleteCascade
      * whether to delete its children when an instance is deleted
+     * @param integer $pageSize
+     * the amount of items to display per page on the listview
      * @param CRUDServiceProvider $serviceProvider
      * The current service provider
      */
-    public function __construct($table, $fields, $label, $listFields, $standardFieldLabels, $childrenLabelFields, $deleteCascade, $serviceProvider) {
+    public function __construct($table, $fields, $label, $listFields, $standardFieldLabels, $childrenLabelFields, $deleteCascade, $pageSize, $serviceProvider) {
         $this->table = $table;
         $this->fields = $fields;
         $this->children = array();
@@ -168,6 +175,7 @@ class CRUDEntityDefinition {
         $this->standardFieldLabels = $standardFieldLabels;
         $this->childrenLabelFields = $childrenLabelFields;
         $this->deleteCascade = $deleteCascade;
+        $this->pageSize = $pageSize;
         $this->serviceProvider = $serviceProvider;
     }
 
@@ -230,6 +238,27 @@ class CRUDEntityDefinition {
      */
     public function setDeleteCascade($deleteCascade) {
         $this->deleteCascade = $deleteCascade;
+    }
+
+    /**
+     * Gets the amount of items to display per page on the listview.
+     *
+     * @return integer
+     * the amount of items to display per page on the listview
+     */
+    public function getPageSize() {
+        return $this->pageSize;
+    }
+
+
+    /**
+    * Sets the amount of items to display per page on the listview.
+    *
+    * @param integer $pageSize
+    * the amount of items to display per page on the listview
+    */
+    public function setPageSize($pageSize) {
+        $this->pageSize = $pageSize;
     }
 
     /**