+++ src/CRUDlex/CRUDServiceProvider.php
@@ -106,6 +106,7 @@ class CRUDServiceProvider implements ServiceProviderInterface {
             $listFields = key_exists('listFields', $crud) ? $crud['listFields'] : null;
             $childrenLabelFields = key_exists('childrenLabelFields', $crud) ? $crud['childrenLabelFields'] : array();
             $deleteCascade = key_exists('deleteCascade', $crud) ? $crud['deleteCascade'] : false;
+            $pageSize = key_exists('pageSize', $crud) ? $crud['pageSize'] : 25;
             $definition = new CRUDEntityDefinition($crud['table'],
                 $crud['fields'],
                 $label,
@@ -113,6 +114,7 @@ class CRUDServiceProvider implements ServiceProviderInterface {
                 $standardFieldLabels,
                 $childrenLabelFields,
                 $deleteCascade,
+                $pageSize,
                 $this);
             $this->datas[$name] = $dataFactory->createData($definition, $fileProcessor);
         }