+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Released: Upcoming
     - "symfony/css-selector": "~2.6",
     - "apigen/apigen": "4.0.0-RC4"
 - Added pagination to the list view
+- Added the ability to hand in default values to the create form
 
 ## 0.9.5
 Released: 2014-11-07